// react-vite-base/src/comps/globals/Anim/enums.ts
export enum TRANSITION_CURVES {
    Linear = "linear",
    Ease = "ease",
    EaseIn = "ease-in",
    EaseOut = "ease-out",
    EaseInOut = "ease-in-out",
    Spring = "cubic-bezier(0.68, -0.55, 0.27, 1.55)",
    Bounce = "cubic-bezier(0.68, -0.6, 0.32, 1.6)",
}

export enum TRANSITIONS {
    Fade = "fade",
    SlideTop = "slide-top",
    SlideBottom = "slide-bottom",
    SlideLeft = "slide-left",
    SlideRight = "slide-right",
    ScaleUp = "scale-up",
    ScaleDown = "scale-down",
    Rotate = "rotate",
    FlipX = "flip-x",
    FlipY = "flip-y",
    ZoomIn = "zoom-in",
    ZoomOut = "zoom-out",
    Bounce = "bounce",
    Shake = "shake",
    Pulse = "pulse",
    Swing = "swing",
    Flip = "flip",
}